tsync: fix video not smooth in vmaster mode [1/1]
PD#SWPL-15541

Problem:
video not smooth during playing dvb demo source.
When tsync in vmaster mode, the pcrscr increase video in every vsync,
no matter whether video buffer is enough or not.

Solution:
When underflow occurs, if tsync in vmaster mode, decrease pcrscr pts.
This will increase the buffer time and make playback smooth

Verify:
AC214

+++ drivers/amlogic/media/frame_sync/timestamp.c
@@ -334,7 +334,7 @@ void timestamp_pcrscr_inc(s32 inc)
 {
 	if (system_time_up) {
 #ifdef MODIFY_TIMESTAMP_INC_WITH_PLL
-		inc = inc * timestamp_inc_factor / PLL_FACTOR;
+		inc = inc * (s32)timestamp_inc_factor / PLL_FACTOR;
 #endif
 		system_time += inc + system_time_inc_adj;
 		ATRACE_COUNTER("PCRSCR",  system_time);
